requirement.txt
langchain==0.3.27
langchain-google-genai==2.1.12
langchain-community==0.3.31
unstructured==0.18.18
faiss-cpu==1.12.0
python-dotenv
streamlit